Fixed command line clash between script path and subcommands.
This causes a noticeable change in behavior. Using scriptisto
with path-expanded script param is no longer possible. E.g.
`scriptisto foo-in-path`, now it must be absolute or relative
path. pathfinder dependency is removed. It just seems too much
logic to embed into the tool. When a script is called via
shebangs the path lookup is done in a shell, so it is not
necessary.